refactor(browser): model copilot runtime entrypoints
