+++ src/gateway/origin-check.test.ts
@@ -1,7 +1,7 @@
 // Browser origin tests document same-origin, private-network, loopback, forwarded
 // host, and explicit allowlist decisions for gateway browser surfaces.
 import { describe, expect, it } from "vitest";
-import { checkBrowserOrigin, isChromeExtensionOrigin } from "./origin-check.js";
+import { checkBrowserOrigin, normalizeChromeExtensionOrigin } from "./origin-check.js";
 
 describe("checkBrowserOrigin", () => {
   it.each([
@@ -175,10 +175,12 @@ describe("checkBrowserOrigin", () => {
   });
 
   it("recognizes only canonical Chrome extension origins", () => {
-    expect(isChromeExtensionOrigin("chrome-extension://abcdefghijklmnopabcdefghijklmnop")).toBe(
-      true,
-    );
-    expect(isChromeExtensionOrigin("chrome-extension://abc")).toBe(false);
-    expect(isChromeExtensionOrigin("https://abcdefghijklmnopabcdefghijklmnop")).toBe(false);
+    expect(
+      normalizeChromeExtensionOrigin("chrome-extension://abcdefghijklmnopabcdefghijklmnop"),
+    ).toBe("chrome-extension://abcdefghijklmnopabcdefghijklmnop");
+    expect(normalizeChromeExtensionOrigin("chrome-extension://abc")).toBeUndefined();
+    expect(
+      normalizeChromeExtensionOrigin("https://abcdefghijklmnopabcdefghijklmnop"),
+    ).toBeUndefined();
   });
 });
